refactor: use single slackEventHandler instead of separating by event type
